+++ provider/aclrule/datasource_rediscloud_acl_rule_read.go
@@ -1,0 +1,70 @@
+package aclrule
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+
+	"github.com/RedisLabs/rediscloud-go-api/redis"
+	"github.com/RedisLabs/rediscloud-go-api/service/access_control_lists/redis_rules"
+
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+func (d *aclRuleDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	// Defensive nil check for client
+	if d.client == nil {
+		resp.Diagnostics.AddError(
+			"Provider Not Configured",
+			"The provider client is not configured. This is an internal error - please report this to the provider developers.",
+		)
+		return
+	}
+
+	var data AclRuleModel
+	diags := req.Config.Get(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	var filters []func(rule *redis_rules.GetRedisRuleResponse) bool
+	filters = append(filters, func(rule *redis_rules.GetRedisRuleResponse) bool {
+		return types.StringValue(redis.StringValue(rule.Name)) == data.Name
+	})
+
+	list, err := d.client.Client.RedisRules.List(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Unable to Read Cloud ACL Rules",
+			fmt.Sprintf("An error occurred while reading ACL rules: %s", err.Error()),
+		)
+		return
+	}
+
+	list = filterRules(list, filters)
+	if len(list) == 0 {
+		diags.AddError(
+			"No ACL Rules Found",
+			"Your query returned no results. Please change your search criteria and try again.",
+		)
+		return
+	}
+
+	if len(list) > 1 {
+		diags.AddError(
+			"Multiple ACL Rules Found",
+			"Your query returned more than one result. Please change try a more specific search criteria and try again.",
+		)
+		return
+	}
+
+	rule := list[0]
+
+	data.ID = types.StringValue(strconv.Itoa(redis.IntValue(rule.ID)))
+	data.Rule = types.StringValue(redis.StringValue(rule.ACL))
+
+	diags = resp.State.Set(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+}