+++ provider/aclrule/testdata/datasource_basic.tf
@@ -1,0 +1,7 @@
+locals {
+  name = "__NAME__"
+}
+
+data "rediscloud_acl_rule" "test" {
+  name = local.name
+}