+++ CHANGELOG.md
@@ -8,6 +8,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ## Fixed
 - `rediscloud_active_active_subscription_database`: Changing the database `name` attribute no longer causes the resource to be destroyed and recreated.
 
+## Changed
+- Migrated `acl_rule` data source from Terraform SDK v2 to the Terraform Plugin Framework. This is an internal architectural change with no breaking changes to the data source schema or behaviour.
+
 # 2.15.0 (28th April 2026)
 
 ## Added